📦  Package the trust score and make it compatible with python 3
