+++ .gitignore
@@ -1,0 +1,7 @@
+build/
+dist/
+*.egg-info/
+
+**/__pycache__
+
+.vscode/
